polish and code docs
